Unify getThinLVInfo and getLVInfo
They produces almost same data exept lv_size and pool_lv. lvm command
can produce both, so unify these functions.
