+++ e2e/snapshot_test.go
@@ -135,9 +135,9 @@ func testSnapRestore() {
 			return err
 		}).Should(Succeed())
 
-		var lv *thinlvinfo
+		var lv *lvinfo
 		Eventually(func() error {
-			lv, err = getThinLVInfo(lvName)
+			lv, err = getLVInfo(lvName)
 			return err
 		}).Should(Succeed())
 
@@ -240,9 +240,9 @@ func testSnapRestore() {
 			return err
 		}).Should(Succeed())
 
-		var lv *thinlvinfo
+		var lv *lvinfo
 		Eventually(func() error {
-			lv, err = getThinLVInfo(lvName)
+			lv, err = getLVInfo(lvName)
 			return err
 		}).Should(Succeed())
 
@@ -270,7 +270,7 @@ func testSnapRestore() {
 		lvName, err = getLVNameOfPVC(restorePVCName, nsSnapTest)
 		Expect(err).Should(Succeed())
 
-		_, err = getThinLVInfo(lvName)
+		_, err = getLVInfo(lvName)
 		Expect(err).Should(Succeed())
 	})
 }
